improve code readability and structure

+++ server/controllers/post.controller.js
@@ -2,6 +2,7 @@ const Post = require("../models/post.model");
 const handleErrors = require("../utils/errors");
 const mongoose = require("mongoose");
 
+// create a new user post using the Post model
 const newPost = async (req, res) => {
   try {
     const post = await Post.create(req.body);
@@ -12,14 +13,19 @@ const newPost = async (req, res) => {
   }
 };
 
+// delete a user post
 const deletePost = async (req, res) => {
   try {
-    const post = await Post.findById(req.params.id);
+    // check if post ID is of ObjectId type
+    const isValid = mongoose.isValidObjectId(req.params.id);
+    if (!isValid) throw Error("invalid post ID");
 
+    // check if the post exists
+    const post = await Post.findById(req.params.id);
     if (!post) throw Error("unavailable");
 
+    // check if user is the owner of the post
     if (req.user.id !== post.userRef) throw Error("unauthorized");
-
     await Post.findByIdAndDelete(req.params.id);
     res.status(200).json("Post deleted");
   } catch (err) {
@@ -28,14 +34,18 @@ const deletePost = async (req, res) => {
   }
 };
 
+// update an existing post
 const editPost = async (req, res) => {
   try {
+    // check if post ID is of ObjectId type
     const isValid = mongoose.isValidObjectId(req.params.id);
     if (!isValid) throw Error("invalid post ID");
 
+    // check if the post exists
     const post = await Post.findById(req.params.id);
     if (!post) throw Error("unavailable");
 
+    // check if user is the owner of the post
     if (req.user.id !== post.userRef) throw Error("unauthorized");
 
     const editedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {
@@ -48,6 +58,7 @@ const editPost = async (req, res) => {
   }
 };
 
+// retrieve a post through its ID
 const getPost = async (req, res) => {
   try {
     const post = await Post.findById(req.params.id);
